Styling and refactoring
