+++ .style.yapf
@@ -1,0 +1,4 @@
+[style]
+based_on_style = yapf
+column_limit = 120
+indent_width = 4
